refactor : 하위 도메인 별로 클래스 분리
