+++ src/main/java/cart/authentication/exception/MemberAuthenticationException.java
@@ -1,4 +1,4 @@
-package cart.authentication;
+package cart.authentication.exception;
 
 public class MemberAuthenticationException extends RuntimeException {
     public MemberAuthenticationException(String message) {
