+++ src/test/java/cart/entity/MemberTest.java
@@ -1,5 +1,6 @@
 package cart.entity;
 
+import cart.authentication.entity.Member;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
