+++ src/test/java/cart/entity/ProductTest.java
@@ -1,5 +1,6 @@
 package cart.entity;
 
+import cart.product.entity.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
