+++ src/test/java/cart/repository/JdbcMemberRepositoryTest.java
@@ -1,8 +1,9 @@
 package cart.repository;
 
-import cart.entity.Member;
-import cart.repository.exception.MemberPersistanceFailedException;
-import org.junit.jupiter.api.Assertions;
+import cart.authentication.repository.JdbcMemberRepository;
+import cart.authentication.entity.Member;
+import cart.authentication.repository.MemberRepository;
+import cart.authentication.exception.MemberPersistanceFailedException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
